move listener middlewares to their own module
