+++ frontend/src/app/store.ts
@@ -1,13 +1,10 @@
 import type { PreloadedState } from "@reduxjs/toolkit";
 import type { Middleware, MiddlewareAPI } from "@reduxjs/toolkit";
-import { combineReducers, configureStore, isAnyOf } from "@reduxjs/toolkit";
+import { combineReducers, configureStore } from "@reduxjs/toolkit";
 import { isRejectedWithValue } from "@reduxjs/toolkit";
 
 import { api } from "@/app/api";
-import {
-  listenerMiddleware,
-  startAppListening,
-} from "@/app/listenerMiddleware";
+import { listenerMiddleware } from "@/app/listenerMiddleware";
 import backendReducer, {
   selectBackendConfigured,
 } from "@/features/backend/backendSlice";
@@ -15,24 +12,11 @@ import cardbacksReducer from "@/features/card/cardbackSlice";
 import finishSettingsReducer from "@/features/finishSettings/finishSettingsSlice";
 import invalidIdentifiersReducer from "@/features/invalidIdentifiers/invalidIdentifiersSlice";
 import modalsReducer from "@/features/modals/modalsSlice";
-import projectReducer, {
-  addMembers,
-  bulkSetQuery,
-  setQuery,
-} from "@/features/project/projectSlice";
-import cardDocumentsReducer, {
-  fetchCardDocumentsAndReportError,
-} from "@/features/search/cardDocumentsSlice";
-import searchResultsReducer, {
-  clearSearchResults,
-} from "@/features/search/searchResultsSlice";
+import projectReducer from "@/features/project/projectSlice";
+import cardDocumentsReducer from "@/features/search/cardDocumentsSlice";
+import searchResultsReducer from "@/features/search/searchResultsSlice";
 import sourceDocumentsReducer from "@/features/search/sourceDocumentsSlice";
-import searchSettingsReducer, {
-  selectSearchSettingsSourcesValid,
-  setFilterSettings,
-  setSearchTypeSettings,
-  setSourceSettings,
-} from "@/features/searchSettings/searchSettingsSlice";
+import searchSettingsReducer from "@/features/searchSettings/searchSettingsSlice";
 import toastsReducer, { setError } from "@/features/toasts/toastsSlice";
 import viewSettingsReducer from "@/features/viewSettings/viewSettingsSlice";
 
@@ -80,34 +64,6 @@ const rtkQueryErrorLogger: Middleware =
 
     return next(action);
   };
-
-startAppListening({
-  matcher: isAnyOf(setSearchTypeSettings, setSourceSettings, setFilterSettings),
-  effect: async (action, { getState, dispatch }) => {
-    /**
-     * Recalculate search results whenever search settings change.
-     */
-
-    const state = getState();
-    const isBackendConfigured = selectBackendConfigured(state);
-    const searchSettingsSourcesValid = selectSearchSettingsSourcesValid(state);
-    if (isBackendConfigured && searchSettingsSourcesValid) {
-      await dispatch(clearSearchResults());
-      await fetchCardDocumentsAndReportError(dispatch);
-    }
-  },
-});
-
-startAppListening({
-  matcher: isAnyOf(addMembers, setQuery, bulkSetQuery),
-  effect: async (action, { dispatch }) => {
-    /**
-     * Fetch card documents whenever new members are added to the project or search results are cleared.
-     */
-
-    await fetchCardDocumentsAndReportError(dispatch);
-  },
-});
 
 //# endregion
